kleaf: Split _ExtModuleInfo from _KernelBuildInfo.
These are the things needed to build kernel_modules /
kernel_module_install, which are external modules.
In other words, these are part of the DDK.

By splitting the providers, this clarifies the interface
that _kernel_build implements to support the DDK. Later,
we can have kernel_filegroup to implement _KernelBuildModuleInfo
as well.

Test: TH
Bug: 218761417
Bug: 211515836

Change-Id: If2d1ce6d48cefa49f633ee70ed555e49123f5619