GKI: ABI: ignore stgdiff symbol type added/removed reports
For reasons as yet unknown, switching between Clang `LTO=thin` and
`LTO=full` results in the loss or gain of type information for some
symbols, in ABI XML as emitted by `abidw`.

```
symbol changed from '__drm_debug' to 'unsigned int __drm_debug'
  type 'unsigned int' was added
```

`stgdiff` reports such changes, but `abidiff`, as we invoke it, does
not.

With this commit such changes are now ignored.

Bug: 221022839
Change-Id: I1c74f37403b786d1d8fa78fdf0fdaf4f242c1149
Signed-off-by: Giuliano Procida <gprocida@google.com>