kleaf: Provide equivalent Bazel command in build.sh
Check-in a script convert_to_bazel.sh that infers an
equivalent bazel command for a given build.sh
command.

Right now, the script finds all _kernel_build targets that
corresponds to the build config same as _setup_env.sh:

  BUILD_CONFIG=${BUILD_CONFIG:-build.config}

... which uses the BUILD_CONFIG env, or the build.config file
(symlink) at the top of repo.

In build.sh, invoke this script and print a banner of the possibly
equivalent Bazel command.

If LTO is specified, an --lto option is also shown in the resulting
command.

To suppress the warning, set
KLEAF_SUPPRESS_BUILD_SH_DEPRECATION_WARNING=1.

Test: manual
Bug: 223403673

Change-Id: I784f4478030de737e9d86e26e3f8dadcaf54b0d7
Signed-off-by: Yifan Hong <elsk@google.com>