kleaf: Initial check-in of script to infer bazel command.
This is the initial check-in of the script convert_to_bazel.sh,
that infers the corresponding bazel command for a given build.sh
command.

Right now, the script finds all _kernel_build targets that
corresponds to the build config same as _setup_env.sh:

  BUILD_CONFIG=${BUILD_CONFIG:-build.config}

... which uses the BUILD_CONFIG env, or the build.config file
(symlink) at the top of repo.

Test: manual
Bug: 223403673

Change-Id: I784f4478030de737e9d86e26e3f8dadcaf54b0d7