kleaf: Add hermetic_tools.
This change adds a hermetic_tools rule that combines
tools from the host and tools from prebuilts, and
provides a path and a setup script for clients to use.

This change also transforms the following:
  //build/kernel:host-tools
  //build/kernel:kernel-build-scripts , the symlinks to prebuilts
to //build/kernel:hermetic-tools, so we have a central place
of tools to build the kernel.

This change also fixes a bug that the kernel build is not
using host tools from Bazel, but rather its own HERMETIC_TOOLCHAIN.
Previously, when _setup_env.sh sees HERMETIC_TOOLCHAIN=1, it
creates its own list of symlinks to the host-tools, making
//build/kernel:host-tools ineffective. This change introduces
a KLEAF_HERMETIC_TOOLCHAIN flag that kernel.bzl passes to
_setup_env.sh, so it disables the creation of $OUT_DIR/host_tools.
The hermetic_tools rule already sets PATH appropriately.

Test: TH
Bug: 222592011

Change-Id: I8d432078faba0c34e3cfa4d14566cb8ca9638e41