kleaf: Add hermetic_tools.
This change adds a hermetic_tools rule that combines
tools from the host and tools from prebuilts, and
provides a path and a setup script for clients to use.

This change also transforms the following:
  //build/kernel:host-tools
  //build/kernel:kernel-build-scripts, things under build-tools
to //build/kernel:hermetic-tools, so we have a central place
of tools to build the kernel.

Test: TH
Bug: 222592011

Change-Id: I8d432078faba0c34e3cfa4d14566cb8ca9638e41