kleaf: copy in-tree modules from module_staging_dir, not out_dir.
If DO_NOT_STRIP_MODULES is empty (meaning that we strip modules),
modules are stripped when they are installed.

Follow the behavior in build.sh that in-tree modules are copied from
module_staging_dir, not out_dir.

This is already the case for external modules, so no need to fix for
kernel_module rule.

Test: TH
Bug: 223912151

Change-Id: Ib89fca3614c1eee345db23139674503489f8c8cd