kleaf: Run workspace status release or non-release builds
For developer builds, it is expected to
run a trivial --workspace_status_command
so the time of determining scmversion does not impact trivial builds.

For that to work, the build servers needs to be configured to
work with --config=release first.

This change trivially adds --config=release to the build.

Test: build with and without --config=release
Bug: 226170550

Change-Id: I1fb946dc2724e6150b9bf30480eaa9ca345b2569