kleaf: Add --config=release.
Add a configuration, `release`, for release builds. Release
builds are (going to be) the configuration for continuous
integration on ci.android.com.

Also add a flag //build/kernel/kleaf:config that is set to
`release` when `--config=release`.

This change also enables the use of --config=release. Without this
change, `bazel build --config=release` will fail because Bazel
does not know about this config.

Bug: 202077908
Test: build with --config=release
Change-Id: I5f17ce4867c2866509d47f971b11537eb32013f9