ANDROID: move ABI diff functionality out of build_abi.sh
This change isolates the build system from the ABI difference checking
and reporting which is now entirely delegated to `diff_abi`.

This should simplify separate maintenance of both the build
system (i.e., the transition to Kleaf) and the ABI monitoring (i.e.,
the transition to `stgdiff`).

With this change, `diff_abi` now

* additionally calls plain `abidiff` in parallel
* emits some extra chat listing all generated reports
* checks multiple exit statuses
* supports --print-report
* prints the first (typically least sensitive) report corresponding to
an ABI break

Bug: 226567130
Change-Id: I52b685ecc10ea1c05a714210b43b05fd097a2165
Signed-off-by: Giuliano Procida <gprocida@google.com>