ANDROID: move ABI diff functionality out of build_abi.sh
This change isolates the build system from the ABI difference checking
and reporting which is now entirely delegated to `diff_abi`.

This should simplify separate maintenance of both the build
system (i.e., the transition to Kleaf) and the ABI monitoring (i.e.,
the transition to `stgdiff`).

With this change, `diff_abi` now

* additionally calls plain `abidiff` in parallel
* lists all generated report paths
* checks all exit statuses
  * reports a break if any of the commands do
* unconditionally prints the first breaking ABI report
  * where reports are ordered from least to most sensitive:
    * the post-processed `abidiff --leaf-changes-only` report
    * the post-processed `abidiff` report
    * the post-processed `stgdiff --format small` report

Bug: 226567130
Change-Id: I52b685ecc10ea1c05a714210b43b05fd097a2165
Signed-off-by: Giuliano Procida <gprocida@google.com>