kleaf: Add git to hermetic tools.
For --strategy KernelBuild=local, scmversion is not pre-calculated
in --workspace_status_command. Instead, it is calculated using
git when the action is executed.
Hence, git is needed for --strategy KernelBuild=local.
However, we do not have ways to select() on --strategy, so just
include git unconditionally for now until we find a better option.

Test: bazel test --strategy KernelBuild=local \
  --strategy KernelConfig=local --config=release \
  //common:kernel_aarch64_test

Fixes: 227131064

Change-Id: I5c6d3c29b8c674dfa3a84a1b743877997ad959e1