kleaf: Optimize kernel_module
... by disassembling the command into multiple parts and
declaring the inputs / outputs properly. The main optimization
is to move the creation of archives (modules_staging_archive
and kernel_uapi_headers_archive) into their own rule, and move
the copying of ko files into its own command too.

In an experiment, this reduces the gs101_soc module from
28s to 21s. It also reduces the build time for all external
kernel modules. Because there are several kernel modules
in the critical path, the savings are very promising.

In the same experiment, the build time for slider_modules_install
reduces from 298s to 260s in the critical path.

Test: TH
Bug: 226673718
Change-Id: I7234468eb2d0457d73ccbebefc81c26bda65ef12
Signed-off-by: Yifan Hong <elsk@google.com>