kleaf: Fix abi_symbollist.raw restore logic for local builds again.
For local builds (--strategy KernelBuild/KernelConfig=local),
we cannot simply hardcode the value
$(relpath abi_symbollist.raw ${KERNEL_DIR}) to UNUSED_KSYMS_WHITELIST
in .config. This is because the .config may be used in
sandboxed actions as well, e.g. when building modules_prepare.

Hence, set it to a more consistent path. In the case of
KERNEL_DIR=common for kernel_aarch64, for example,
the config is set to:

  ../bazel-out/k8-fastbuild/bin/common/kernel_aarch64_config_intermediates/abi_symbollist.raw

Then, in the setup script that kernel_config provides, do
the following:

- Calculate abs_srctree=$(realpath $ROOT_DIR/$KERNEL_DIR), just like
  common/Makefile
- Calculate CONFIG_UNUSED_KSYMS_WHITELIST by reading .config.
- Restore to $abs_srctree/$CONFIG_UNUSED_KSYMS_WHITELIST.

In the case of sandbox builds, $abs_srctree is within the
execroot/sandbox, so restoring it won't have side effects.

In the case of local builds, $abs_srctree points to the real
source tree (<workspace_root>/common), so this resolves to

  <workspace_root>/bazel-out/..._intermediates/abi_symbollist.raw

... which technically isn't the intermediates directory for the
KernelBuild action. But since we use rsync and preserve all
modes / timestamps / ownership etc., and the file's content should be
consistent throughout the whole Bazel command, this is acceptable.

Test: TH
Test: manually build //common:kernel_aarch64_dist with \
   --strategy KernelConfig=local --strategy KernelBuild=local
Fixes: 228528064
Fixes: c263c4b0ecf788952bca4cda3f5364cd42d4f426
  (kleaf: Do not restore abi_symbollist.raw to KERNEL_DIR for non-sandbox builds.)

Change-Id: I6f5ba8cd6af1b0b3954b13cfe909c0981cd1b868