ANDROID: do not trust plain abidiff report
The plain `abidiff` report sometimes contains ABI diffs which are not
reported by `abidiff --leaf-changes-only`.

It's highly likely that these diffs are due to only declared vs fully
defined type issues which we are ignoring for pragmatic reasons.

Given that we are moving away from using `abidiff` altogether, it
makes sense to ignore the outcome of plain `abidiff`. The report is
still generated unconditionally.

Bug: 227282653
Change-Id: I3a0fefb89df80341c66cc13e95f8b0bbeed60d2f
Signed-off-by: Giuliano Procida <gprocida@google.com>