kleaf: set scmversion and source_date_epoch only for config=release.
Only when --config=release is set, set scmversion
and source_date_epoch.

A side effect of this change is that --config=local
now uses --workspace_status_command=(empty).
When --config=local, some actions run in the sandbox and some
does not. The previous behavior causes actions to be rebuilt
because ctx.info_file is changed even though the corresponding
variable does not change. This change reduces unintentional
rebuilds.

A side effect of this change is to remove the harmless warning on
EXT_MODULES for GKI builds, where the top level build.config does not exist.

Bug: 228689043
Test: build slider with and without --config=release

Change-Id: I3af7a61cea71dd0fcf5763e68bcb5e717931dd70