kleaf: skip scmversion and source_date_epoch for --config=skip_scm.
When --config=skip_scm is set, skip scmversion
and source_date_epoch.

--config=local implies --config=skip_scm.
When --config=local, some actions run in the sandbox and some
does not. The previous behavior causes actions to be rebuilt
because ctx.info_file is changed even though the corresponding
variable does not change. This change reduces unintentional
rebuilds.

--config=fast implies --config=local which implies --config=skip_scm.
The fast config is recommended on local builds.

A side effect of this change is to remove the harmless warning on
EXT_MODULES for GKI builds when --config=skip_scm,
where the top level build.config does not exist.

Bug: 228689043
Test: build slider with and without --config=release

Change-Id: I3af7a61cea71dd0fcf5763e68bcb5e717931dd70