ANDROID: GKI: abitool.py: always create short reports
This change mades `diff_abi` unconditionally create short reports and
removes the corresponding flag passing from `build_abi.sh`. There is
no change in behaviour for callers of `build_abi.sh`.

Callers of `diff_abi` that do not pass a `--short-report` argument
will get an additional short report regardless.

This is needed because any ABI break overriding via dropping of diff
paragraphs happens during postprocessing of ABI reports and the only
reports that are the result of post-processing are the short reports.

Change-Id: Ibad31b9b8a755ca264093e89f9ebcda1fbf51ff6
Signed-off-by: Giuliano Procida <gprocida@google.com>