ANDROID: GKI: drop ignored lines from short STG ABI diff report
The short STG report is generated from the small STG report.

If the small report only contained ignorable paragraphs, this was
treated as no ABI break. This commit modifies the behaviour to drop
the ignorable paragraphs during creation of the short report and
override ABI break status iff the resulting report is empty.

There should be no change in behaviour in terms of reporting the
presence of an ABI break but the short report no longer contains the
ignored paragraphs.

This commit folds _reinterpret_stgdiff into _shorten_stgdiff and adds
_remove_matching_lines as a helper.

Change-Id: I2ac8efda489aa3deef076433f88473725d2acbba
Signed-off-by: Giuliano Procida <gprocida@google.com>