ANDROID: GKI: drop ignored lines from short STG ABI diff report
The short STG report is generated from the small STG report.

If the small report only contained ignorable paragraphs, the exit code
was overridden to 0. This commit modifies the behaviour to dropping
the ignorable paragraphs when creating the short report and overriding
the exit status if the resulting report is empty.

There should be no change in behaviour in terms of reporting the
presence of an ABI break but the short report no longer contains the
ignored paragraphs.

Change-Id: I2ac8efda489aa3deef076433f88473725d2acbba
Signed-off-by: Giuliano Procida <gprocida@google.com>