kleaf: _notrim and _with_vmlinux targets should not define base_kernel & add System.map
For ABI monitoring purposes, we don't care about
mixed builds, because it needs to build a vmlinux
anyways.

It is incorrect because it tries to depend on
base_kernel, which may be //common:kernel_aarch64,
which we don't know if vmlinux is trimmed or not.
This makes the _notrim target requiring the GKI
(possibly trimmed) vmlinux to be built, which may fail.

Specifying the "default" argument for dict.pop allows
it to not fail when the key is not found.

Now that we don't have base_kernel, make sure both vmlinux
and System.map is built so that external modules don't fail
with:

```
  INFO: From Building external kernel module //gs/google-modules/touch/common:common.slider_notrim:
  Warning: modules_install: missing 'System.map' file. Skipping depmod.
```

Test: TH
Bug: 231480193

Change-Id: Ie4a89a5426c375c4ed4400579060d5f3b943e15b