build.sh: Expand $GKI_MODULES_LIST to abs path
It is not a good style to assume a current working directory
and to accommodate the relative paths against it. Better and
more reliable solution is to use absolute paths, otherwise
there is a risk to hit a case where it is different, hence
relative path for $GKI_MODULES_LIST will not exist.

Change-Id: Ica14531be12d30610039319f2bbe915b4990a3aa
Signed-off-by: Blagovest Kolenichev <quic_c_bkolen@quicinc.com>