download_from_ci: get_binary_kernel_version works for arm64 and x86
So far, get_binary_kernel_version worked only for arm64 and not for
x86_64. This was because the function was implemented by grep'ing a
magic string in the kernel binary. The kernel is compressed (i.e.
bzImage) for x86_64.

This change fixes the problem by using file(1) which can handle both
uncompressed Image and compressed bzImage.

It falls back to the legacy grep-path when file(1) doesn't show the
version string which currently is the case for arm64.

Bug: N/A
Test: run download_from_ci
Change-Id: I05f19a78377a6a07ba96ee894a0796af461f56df