Kleaf: fix caching for --config=local
The --config=local mode was broken in that it would not actually cache
the OUT_DIR across executions. This was due to the OUT_DIR being defined
in the sandbox, relative to symlinked sources. It is not clear why the
local mode worked before and what combination of changes broke it. This
change restores is by adding a starlark flag --cache_dir that is used to
define OUT_DIR in KernelEnv. Said flag is also understood by the
bazel.py wrapper and set to a reasonable default within WORKSPACE if not
defined. Yet in order to have $WORKSPACE/out/cache we need to also
execute the KernelEnv rule in config=local mode.

This lets --config=local and --config=fast cache OUT_DIR in
$WORKSPACE/out/cached/<some_generated_name> across invocations. Further,
by specifying --cache_dir=/some/fast/disk consistently across
invocations, one can make use of a fast filesystem for caching; perhaps
a tmpfs.

In an experiment, this saves 50s when comparing --config=local
and non --config=local incremental builds.

Bug: 235632059
Test: manual with the following:
  $ bazel build //common:kernel_dist                && echo >> common/init/main.c && time bazel build //common:kernel_dist
  Takes 114 seconds for the second build; kernel_build() takes 97s.

  $ bazel build --config=local //common:kernel_dist && echo >> common/init/main.c && time bazel build --config=local //common:kernel_dist
  Takes 64  seconds for the second build; kernel_build() takes 40s.

Signed-off-by: Matthias Maennich <maennich@google.com>
Signed-off-by: Yifan Hong <elsk@google.com>
Change-Id: I64fbfd928efd78fb5d3d3c3af6a5582fca22b0a5
(cherry picked from commit c2f2105abc6f729357a56a6c354ac9516a4a0907)