build_abi.sh: fix symbol list updating for mixed builds
When updating the ABI symbol list for a mixed build configuration,
build_abi.sh disables mixed building and directly compiles the kernel
using the device BUILD_CONFIG. In order to be able to successfully do
this, we need to also make sure the proper kernel files are copied over
to the DIST_DIR (since mixed building doesn't require the device
BUILD_CONFIG to do this). Re-works this logic to (1) only modify
MAKE_GOALS and FILES when mixed building is enabled and (2) append the
necessary kernel artifacts to the FILES variable.

Bug: 239094500
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: If1f1d60874ba899f4f8e50394b9eb653a1f7cef9