kleaf: Add build_config_to_bazel.py
This is a helper script that helps the transition
from build.sh build configs to Bazel rules.

Sample usage:
  BUILD_CONFIG=common-modules/virtual-device/build.config.virtual_device.aarch64 build/kernel/kleaf/build_config_to_bazel.py
This will create/change common-modules/virtual-device/BUILD.bazel
to add / modify rules to build the virtual device target. The
generated rules are not buildable immediately; manual intervention
is required with FIXME comments in the generated file.

Test: build/kernel/kleaf/build_config_to_bazel_test.py
Bug: 241320850
Change-Id: I8fa6eb49c1c21fcfa7f1e972f81a8a248000b0f6