kernel/abi: remove gki_check script
gki_check is outdated and doesn't support GKI kernels beyond 5.4
There are no plans to maintain the tool forward as abi tooling is
handling it's purpose better way at this point. The doc reference
have been removed in the past via cl/446725767; and GKI newsletter
announcement for the removal of the script was published on 08/12/22
with 2 weeks notice.

Bug: 230765662
Test: TH
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>
Change-Id: I5d8b3ae465b00e3c394b287457403015348a2dbc