Add toybox' zcat to PATH
zcat is needed for cases when tar's autodetection for the filetype kicks
in, e.g. `tar xf file.tar.gz`. While the detection is working, a
detection for zcat in the PATH let's toybox fall back to bzip2 as a
default tool, which is a bad choice as it will not even fail, but just
emit `bzip2: (stdin) is not a bzip2 file.` to stdout.

Bug: 243570975
Bug: 169934030
Change-Id: I7abf34f66b1bbab65c1d56b8d4a3de9646031c3a