kleaf: Do not include GKI modules in default output of //common:kernel_aarch64.
When a device kernel builds their own modules that
conflicts with the GKI modules (zram and zsmalloc
for android13-5.15), the dist target will produce a
warning about conflicting source files.

Hence, do not include zram and zsmalloc to the device
dist targets by default when they include "//common:kernel_aarch64"
only.

If they wish to include zram and zsmalloc in
the distribution on the android13-5.15 branch,
they may add the following to the
dist target:

  "//common:kernel_aarch64/zram.ko",
  "//common:kernel_aarch64/zsmalloc.ko",

//common:kernel_aarch64_dist continues to include GKI
modules defined by the branch.

Bug: 244215515
Test: manual

Change-Id: Id1c9fd9e603e9b2946cd59c7fde14d11ef7dd8b2