kleaf: Remove toolchain_version from the default output of kernel_build.
The file was added to DefaultInfo of kernel_build
only to allow downloading toolchain_version from
build servers so that kernel_filegroup works. However,
this caused a conflict when the device's kernel_build
and the GKI (//common:kernel_aarch64) are both added
to a copy_to_dist_dir rule with flat=True (even though
the two files should have the same content).

Hence, drop it from the DefaultInfo. To continue to download
it from build servers, it is kept in //common:kernel_aarch64_dist.

Test: TH
Bug: 244619402

Change-Id: I0e69c5bff6207f75ac41a440e33958646fdea926