extract_symbol: symbol list update only consider vendor modules
When an inbuilt kernel feature gets converted to a module,
the tracked KMI (i.e. the surface of the ABI we keep stable)
should be unchanged as newly added symbols are not trimmed to
allow loading the GKI modules. To meet this requirement only
consider the symbols from local i.e. vendor modules for the
symbol list update purposes.

Bug: 232431411
Test: TH
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>
Change-Id: I536bf8eec311d26aefdf79f0c9d68c53c257a9c5