abi: Enable renumbering of anonymous types ids by abitidy
This change will instruct abitidy to renumber anonymous type ids using
the information about the first member of the type for anonymous structs
or unions, or the first enumerator for anonymous enums.

The above tranformation makes sure that all anonymous types of the same
kind do not hash to the same ids, hence reducing the need for linear
probing during type id generation and avoid shifts in large number of
anonymous type ids when new anonymous types are added to the ABI
represenation.

Bug: 239910948
Signed-off-by: Sid Nayyar <sidnayyar@google.com>
Change-Id: I442c9f0495202266ec89193553888194e5d91296