Revert "kleaf: refactor get_include_depset and get_headers_depset"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/5ee83909-df65-4953-a816-9d5626362dcd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6057963581a8c8d7a0177a60d4bbd56d938de597