kleaf: extract_symbol with GKI modules objects
To partition between GKI and vendor modules on the fly;
we need to pass GKI modules from the base_kernel build
in addition to vendor modules. Remove passing the list
and pass the modules instead from the base_kernel
first and then the vendor modules overwriting the modules
from the base build being overridden.

Bug: 232431411
Test: TH
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>
Change-Id: I2f40943d0d4f0a65cc7ab47a5064caff7fbf39b4